// src/features/editorGarlia/components/types.ts
export type Reino = {
  id: string;
  nombre: string;
  historia?: string;
  politica?: string;
  economia?: string;
  geografia?: string;
  cultura?: string;
  mapa_url?: string;
  coord_x?: number;
  coord_y?: number;
  oculto?: boolean;
};
// ...
export type SaveStatus = "idle" | "saving" | "saved" | "error";